Cart Service Custom exceptions implemented successfully

+++ src/main/java/service/CartService.java
@@ -2,6 +2,9 @@ package service;
 
 import dao.CartDAO;
 import dao.CartItemDAO;
+import exception.CartEmptyException;
+import exception.CartItemUnavailableException;
+import exception.InvalidQuantityException;
 import model.Cart;
 import model.CartItem;
 
@@ -19,6 +22,10 @@ public class CartService {
 	}
 
 	public void addToCart(int userId, int productId, int quantity) {
+        if(quantity<=0)
+        {
+            throw new InvalidQuantityException("Quantity must be greater than 0");
+        }
 		Cart cart = cartDAO.getCartByUserId(userId);
 		if (cart == null) {
 			int cartId = cartDAO.createCart(userId);
@@ -35,33 +42,37 @@ public class CartService {
 		}
 	}
 
-	public void viewCart(int userId) {
+	public List<CartItem> viewCart(int userId) {
 		Cart cart = cartDAO.getCartByUserId(userId);
 		if (cart == null) {
-			System.out.println("Cart in Empty!!! ");
-            return;
+			throw new CartEmptyException("Cart is empty");
 		}
 
         List<CartItem> items= cartItemDAO.getCartItems(cart.getCartId());
         if(items.isEmpty())
         {
-            System.out.println("Cart is empty");
-            return;
+            throw new CartEmptyException("Cart is empty");
         }
-        System.out.println("==== Cart Items ====");
-        for(CartItem item:items)
-        {
-            System.out.println("Product id: "+item.getProductId()  + "====" + "Quantity:    "+item.getQuantity());
-        }
+        return items;
+//        System.out.println("==== Cart Items ====");
+//        for(CartItem item:items)
+//        {
+//            System.out.println("Product id: "+item.getProductId()  + "====" + "Quantity:    "+item.getQuantity());
+//        }
 	}
 
 
     public List<CartItem> getCartItemsByUserId(int userId) {
         Cart cart = cartDAO.getCartByUserId(userId);
         if (cart == null) {
-            return new ArrayList<>();
+           throw new CartEmptyException("Cart is empty");
         }
-        return cartItemDAO.getCartItems(cart.getCartId());
+        List<CartItem> items= cartItemDAO.getCartItems(cart.getCartId());
+        if(items.isEmpty())
+        {
+            throw new CartEmptyException("Cart is empty");
+        }
+        return items;
     }
 
     public void removeFromCart(int userId,int productId)
@@ -69,8 +80,12 @@ public class CartService {
         Cart cart= cartDAO.getCartByUserId(userId);
         if(cart==null)
         {
-            System.out.println("Cart doesn't exist");
-            return;
+           throw new CartItemUnavailableException("Item not found in cart with user id"+userId +"and" +"product id"+productId);
+        }
+        CartItem cartItem = cartItemDAO.getCartItem(cart.getCartId(), productId);
+        if(cartItem == null)
+        {
+            throw new CartItemUnavailableException("Item not found in cart with user id"+userId +"and" +"product id"+productId);
         }
         cartItemDAO.removeItem(cart.getCartId(),productId);
